+++ MyrranMP/src/DTO/NetDTO.java
@@ -138,9 +138,9 @@ public class NetDTO
         public int celdaX;
         public int celdaY;
         public int numCapa;
-        public int iDTerreno;
+        public short iDTerreno;
         public SetTerreno() {}
-        public SetTerreno(int celdaX, int celdaY, int numCapa, int iDTerreno)
+        public SetTerreno(int celdaX, int celdaY, int numCapa, short iDTerreno)
         {   this.celdaX = celdaX; this.celdaY = celdaY; this.numCapa = numCapa; this.iDTerreno = iDTerreno; }
     }
